Fix AT+VTS=n DTMF tones for BT handsfree on CDMA networks.
When calling voice mail or bank using a BT carkit with numeric keypad.
When pressing 0-9,#,* in a voice menu the AT+VTS=n command is sent
from BT carkit to the device. For CDMA no DTMF are sent over the network
and no tones are generated. For UMTS it sends a correct DTMF command,
but still no tones are generated.
For CDMA phone the AT+VTS=n command handler should send correct
DTMF commands to the network by using correct method. And
for both phones CDMA and UMTS we should generate local DTMF tones

Change-Id: Ib461cafa50cc53d2940e118b5c6751ef8affe0d9